+++ crates/forge/src/run_tests/workspace.rs
@@ -1,5 +1,6 @@
 use super::package::RunForPackageArgs;
 use crate::profile_validation::check_compiler_config_compatibility;
+use crate::profile_validation::enable_gas::check_enable_gas;
 use crate::run_tests::messages::latest_blocks_numbers::LatestBlocksNumbersMessage;
 use crate::run_tests::messages::overall_summary::OverallSummaryMessage;
 use crate::run_tests::messages::partition::{PartitionFinishedMessage, PartitionStartedMessage};
@@ -14,8 +15,9 @@ use crate::{
 };
 use anyhow::Result;
 use forge_runner::partition::PartitionConfig;
+use forge_runner::resolve_cache_dir;
 use forge_runner::test_case_summary::AnyTestCaseSummary;
-use forge_runner::{CACHE_DIR, test_target_summary::TestTargetSummary};
+use forge_runner::test_target_summary::TestTargetSummary;
 use foundry_ui::UI;
 use scarb_api::metadata::{MetadataOpts, metadata_with_opts};
 use scarb_api::{
@@ -42,6 +44,8 @@ pub async fn run_for_workspace(args: TestArgs, ui: Arc<UI>) -> Result<ExitStatus
 
     let packages: Vec<PackageMetadata> =
         packages_from_filter(&scarb_metadata, &args.scarb_args.packages_filter)?;
+
+    check_enable_gas(&scarb_metadata, &packages)?;
 
     let filter = PackagesFilter::generate_for::<Metadata>(packages.iter());
 
@@ -98,11 +102,10 @@ pub async fn execute_workspace(
 
     let block_number_map = BlockNumberMap::default();
     let mut all_tests = vec![];
-    let mut total_filtered_count = Some(0);
+    let mut total_filtered_count = 0;
     let mut exit_first_channel = ExitFirstChannel::new();
 
-    let workspace_root = &scarb_metadata.workspace.root;
-    let cache_dir = workspace_root.join(CACHE_DIR);
+    let cache_dir = resolve_cache_dir(&scarb_metadata.workspace.root)?;
     let packages_len = packages.len();
 
     let partitioning_config = get_partitioning_config(args, &ui, &packages, &artifacts_dir_path)?;
@@ -138,9 +141,8 @@ pub async fn execute_workspace(
         )
         .await?;
 
-        let filtered = result.filtered();
+        total_filtered_count += result.filtered();
         all_tests.extend(result.summaries());
-        total_filtered_count = calculate_total_filtered_count(total_filtered_count, filtered);
         env::set_current_dir(&cwd)?;
     }
 
@@ -185,18 +187,6 @@ pub async fn execute_workspace(
     }
 
     Ok(WorkspaceExecutionSummary { all_tests })
-}
-
-fn calculate_total_filtered_count(
-    total_filtered_count: Option<usize>,
-    filtered: Option<usize>,
-) -> Option<usize> {
-    // Calculate filtered test counts across packages. When using `--exact` flag,
-    // `result.filtered_count` is None, so `total_filtered_count` becomes None too.
-    match (total_filtered_count, filtered) {
-        (Some(total), Some(f)) => Some(total + f),
-        _ => None,
-    }
 }
 
 fn get_partitioning_config(