+++ crates/sncast/src/starknet_commands/account/import.rs
@@ -1,14 +1,13 @@
 use std::str::FromStr;
 
-use super::deploy::compute_account_address;
 use crate::starknet_commands::account::{
-    generate_add_profile_message, prepare_account_json, write_account_to_accounts_file,
+    compute_account_address, generate_add_profile_message, prepare_account_json,
+    write_account_to_accounts_file,
 };
 use anyhow::{Context, Result, bail, ensure};
 use camino::Utf8PathBuf;
 use clap::Args;
 use conversions::string::{TryFromDecStr, TryFromHexStr};
-use foundry_ui::components::warning::WarningMessage;
 use sncast::check_if_legacy_contract;
 use sncast::helpers::account::generate_account_name;
 use sncast::helpers::configuration::CastConfig;
@@ -87,14 +86,6 @@ pub async fn import(
     config: &CastConfig,
     ui: &UI,
 ) -> Result<AccountImportResponse> {
-    // TODO(#3556): Remove this warning once we drop Argent account type
-    if import.account_type == AccountType::Argent {
-        ui.print_warning(WarningMessage::new(
-                "Argent has rebranded as Ready. The `argent` option for the `--type` flag in `account import` is deprecated, please use `ready` instead.",
-            ));
-        ui.print_blank_line();
-    }
-
     let (signer_type, public_key) = if let Some(ledger_path) = import.ledger_key_locator.resolve(ui)
     {
         let public_key = ledger::get_ledger_public_key(&ledger_path, ui).await?;
@@ -156,8 +147,16 @@ pub async fn import(
     let chain_id = get_chain_id(provider).await?;
 
     if let Some(salt) = import.salt {
-        let computed_address =
-            compute_account_address(salt, public_key, class_hash, import.account_type, chain_id);
+        let computed_address = compute_account_address(
+            salt,
+            class_hash,
+            import.account_type,
+            chain_id,
+            &signer_type,
+            provider,
+            ui,
+        )
+        .await?;
         ensure!(
             computed_address == import.address,
             "Computed address {:#x} does not match the provided address {:#x}. Please ensure that the provided salt, class hash, and account type are correct.",