+++ crates/sncast/tests/e2e/main_tests.rs
@@ -7,8 +7,8 @@ use crate::helpers::fixtures::{
 };
 use crate::helpers::runner::runner;
 use configuration::test_utils::copy_config_to_tempdir;
-use indoc::indoc;
-use shared::test_utils::output_assert::assert_stderr_contains;
+use indoc::{formatdoc, indoc};
+use shared::test_utils::output_assert::{assert_stderr_contains, assert_stdout_contains};
 
 #[tokio::test]
 async fn test_happy_case_from_sncast_config() {
@@ -223,6 +223,28 @@ async fn test_missing_account_flag() {
 }
 
 #[tokio::test]
+async fn test_missing_account_flag_json() {
+    let args = vec![
+        "--json",
+        "--accounts-file",
+        ACCOUNT_FILE_PATH,
+        "declare",
+        "--url",
+        URL,
+        "--contract-name",
+        "whatever",
+    ];
+
+    let snapbox = runner(&args);
+    let output = snapbox.assert().failure();
+
+    assert_stderr_contains(
+        output,
+        r#"{"command":"declare","error":"Account name not passed nor found in snfoundry.toml","type":"error"}"#,
+    );
+}
+
+#[tokio::test]
 async fn test_inexistent_keystore() {
     let accounts_file = "empty_accounts.json";
     let temp_dir = tempfile::tempdir().expect("Unable to create a temporary directory");
@@ -384,7 +406,15 @@ async fn test_keystore_declare() {
     set_keystore_password_env();
     let snapbox = runner(&args).current_dir(contract_path.path());
 
-    assert!(snapbox.assert().success().get_output().stderr.is_empty());
+    let output = snapbox.assert().success();
+    assert!(output.get_output().stderr.is_empty());
+    assert_stdout_contains(
+        output,
+        formatdoc! {"
+            To deploy a contract of this class, run:
+            sncast --keystore {my_key_path} --account {my_account_path} deploy --class-hash 0x[..] --url [..]
+        "},
+    );
 }
 
 #[tokio::test]