+++ crates/snforge-scarb-plugin/src/attributes/fuzzer/wrapper.rs
@@ -120,8 +120,8 @@ fn fuzzer_wrapper_internal(
 
                     #func_name(#blank_values_for_config_run);
 
-                    return;
-                }
+                    return Default::default();
+                };
                 #fuzzer_assignments
                 #func_name(#arguments_list);
             }